Suggestions for increased transparency
This PR suggests some text changes to the copy, and adds back a chronology.

Because our site is used by:

1. Auditors and auditees submitting audits
2. State auditors (and other entities) responsible for passthrough monitoring
3. Federal audit resolution officials
4. IG offices
5. Grants managers/grants issuing offices
6. GAO
8. OMB
9. The general public
10. ...

I think credibility is enhanced by transparency, and because we have a
public record of our efforts in this space (our Github repository), I
think we should use it to document the actions we've taken so far.

For consideration.