Starting on transferring data to HTML
1. Everything needs to be YAML or JSON.
2. I've created a stub data file with the beginnings of the data for
   each page.
3. I've created a page that starts to render out the contents for each
   type of data.

Note that the KSAMLs are actually complete. But, because they were a CSV
file, we have to convert it to a JSON document first. I have a bash
script that does that. Someday, it would be nice if that was part of the
build process...

This branch serves as a starting point.

We may also want a new page that `helpdesk` goes to; that page would
help people understand where to go for what questions, with links to
each of the above pages as appropriate.